requirements.txt
streamlit==1.37.0
python-dotenv==1.0.1
google-generativeai==0.6.10